Instructions to run new code:
No paths are hardcoded instead raw data must be placed within a folder named "raw". Inside the directory in which this file lies, the processed data will be saved with the same path, but in a new "processed" folder.

Runs containing multiple folders may be processed with the -d option (multi-threading not implemented yet)