Only one TTree per TFile, fixed memory leaks
Added an option to the tree write command such that only the most
recent revision to a TTree is written to disk.  Also fixed some memory
leaks caused byT Files not being deleted.  Added an option to the tree
write command such that only the most recent revision to a TTree is
written to disk.  Also fixed some memory leaks caused by TFiles not
being deleted.